+++ zerver/lib/events.py
@@ -34,14 +34,15 @@ from zerver.lib.actions import (
     do_get_streams, get_default_streams_for_realm,
     gather_subscriptions_helper, get_cross_realm_dicts,
     get_status_dict, streams_to_dicts_sorted,
-    default_stream_groups_to_dicts_sorted
+    default_stream_groups_to_dicts_sorted,
+    get_owned_bot_dicts,
 )
 from zerver.lib.user_groups import user_groups_in_realm_serialized
 from zerver.tornado.event_queue import request_event_queue, get_user_events
 from zerver.models import Client, Message, Realm, UserPresence, UserProfile, \
     get_user_profile_by_id, \
     get_realm_user_dicts, realm_filters_for_realm, get_user,\
-    get_owned_bot_dicts, custom_profile_fields_for_realm, get_realm_domains, \
+    custom_profile_fields_for_realm, get_realm_domains, \
     get_default_stream_groups
 from zproject.backends import email_auth_enabled, password_auth_enabled
 from version import ZULIP_VERSION
